split parseReferenceObject function to a file
