+++ lib/v3.1/schema/parse.js
@@ -5,6 +5,9 @@ const {
   MEDIA_TYPE,
   PARAMETER_LOCATION
 } = require('../constants')
+const {
+  parseReferenceObject
+} = require('../parseReferenceObject')
 
 // https://datatracker.ietf.org/doc/html/draft-bhutton-json-schema-00#section-4.2.1
 function valueToLiteral (value) {
@@ -29,44 +32,6 @@ function valueToLiteral (value) {
   }
 
   throw new Error(`unexpected default value type. [${typeof value}]`)
-}
-
-function resolveReferenceObject (
-  inputPath,
-  openapiObject,
-  paths,
-  currentObject
-) {
-  if (paths.length > 1) {
-    return resolveReferenceObject(
-      inputPath,
-      openapiObject,
-      paths,
-      currentObject[paths.shift()]
-    )
-  }
-
-  return currentObject[paths.shift()]
-}
-
-// https://github.com/OAI/OpenAPI-Specification/blob/main/versions/3.1.0.md#reference-object
-function parseReferenceObject (
-  inputPath,
-  openapiObject,
-  reference
-) {
-  if (typeof reference !== 'string') {
-    // https://github.com/OAI/OpenAPI-Specification/blob/main/versions/3.1.0.md#reference-object
-    // > REQUIRED. The reference identifier. This MUST be in the form of a URI.
-    throw new Error(`The reference identifier is must be in the form of a URI. Current value: ${reference}`)
-  }
-
-  return resolveReferenceObject(
-    inputPath,
-    openapiObject,
-    reference.split('/').slice(1),
-    openapiObject
-  )
 }
 
 // https://github.com/OAI/OpenAPI-Specification/blob/main/versions/3.1.0.md#schema-object
